planner-adk/requirement.txt
fastapi[standard]
google-adk
uvicorn
python-dotenv
google-cloud-pubsub
httpx
pydantic
